requirements.txt
attrs==16.3.0
Automat==0.5.0
constantly==15.1.0
Django==1.11
incremental==16.10.1
pytz==2017.2
six==1.10.0
Twisted==17.1.0
zope.interface==4.3.3
south
djangorestframework
autopep8
pycodestyle
decorator
requests

# Pylint
pylint-django
pylint
colorize

#Extractor
dpkt
netaddr
GeoIP
user_agents

elasticsearch>=5.0.0,<6.0.0
httpagentparser
bs4
click
